Ic 33051

+++ scripts/Python/MCNP/clean_output.py
@@ -2,14 +2,19 @@ import os
 import logging
 
 def cleanOutputFolder():
-    os.chdir("output/")
-    all_files = os.listdir()
 
-    for f in all_files:
-        os.remove(f)
+    if os.path.exists("output/"):
 
-    logging.info("All files in output_nps10E7 folder were removed")
+        print('Folder exists!')
+        os.chdir("output/")
+        all_files = os.listdir()
 
+        for f in all_files:
+            os.remove(f)
+
+        logging.info("All files in output_nps10E7 folder were removed")
+    else:
+        print('Folder does not exist.')
 
 if __name__ == '__main__':
     cleanOutputFolder()
